Fix compile options and some warnings
Former-commit-id: 8758ce1ee9d46005e0d8cde63e65bba48d559474 [formerly e9a98962c6dd66319d688c00730419066ea452f7] [formerly 849b652812c20c4b4f080b2e3188c0fc3944fd86] [formerly 4ee4ba66f1b6f9305b122bae376fbe116747f6ca [formerly 58f8699daa17e2c1c5024b00ad10dc81a8d86c08] [formerly 849b652812c20c4b4f080b2e3188c0fc3944fd86 [formerly 849b652812c20c4b4f080b2e3188c0fc3944fd86 [formerly 954c1dd9f19d2bdaf77a4cacd92a6ff9be159254]]]]
Former-commit-id: 776803d6c22cb771ebc1339f5e417de6d8cb7c7d [formerly 0312b4922ac4c8482bae3eba89b43f1a85991e62] [formerly 8585abc55838c592c3e9e55d16673575c0589cc4 [formerly 1dae94f6ab8ed521914fa035966d67feddce0b8e]]
Former-commit-id: 8585abc55838c592c3e9e55d16673575c0589cc4
Former-commit-id: 9f70b6896fb022411b184b98175e937d862be817
Former-commit-id: 2e7b9975fd85980f4b9c1dd5586a505240d0104f

+++ src/Poulpe/Component/Mesh.cpp
@@ -62,7 +62,7 @@ namespace Poulpe
             float yMin = yMax;
             float zMin = zMax;
 
-            for (int j = 0; j < data.m_Vertices.size(); j++) {
+            for (size_t j = 0; j < data.m_Vertices.size(); j++) {
 
                 glm::vec3 vertex = glm::vec4(data.m_Vertices.at(j).pos, 1.0f);
